Skip file when key file is missing instead of aborting initialization
We have observed instances where .key files are missing but the main file still
exists.  It's not entirely clear how this happens (file system corruption?),
but in case it does, it makes sense just to delete the relevant file and
continue, rather than aborting the entire initialization as happens currently.