fix #17: proper embedded alignment of handleCounter
amd32 (raspberrypi) requires specific alignment for atomic operations to
work, or else you'll get "panic: runtime error: invalid memory address
or nil pointer dereference". Turns out this alignment is important for
embedded structs as well.
Embedding as the first member in the struct, and putting the atomic
field at the start of the embedded struct makes the alignment correct.